backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-mptt==0.16.0
python-decouple==3.8
gunicorn==21.2.0
